style: polish Vaultly UI to match other apps
